+++ src/app/shared/services/icon.service.ts
@@ -86,6 +86,13 @@ export class IconService {
       webdav: 'assets/icons/remotes/webdav.svg',
       sftp: 'assets/icons/remotes/sftp.svg',
       hdfs: 'assets/icons/remotes/hdfs.svg',
+      box: 'assets/icons/remotes/box.svg',
+      yandex: 'assets/icons/remotes/yandex.svg',
+      'google photos': 'assets/icons/remotes/google.svg',
+      'google cloud storage': 'assets/icons/remotes/google.svg',
+      jottacloud: 'assets/icons/remotes/jottacloud.svg',
+      hidrive: 'assets/icons/remotes/hidrive.svg',
+      pcloud: 'assets/icons/remotes/pcloud.svg',
 
       // ------------------- Theme Icons -------------------
       'circle-check': 'assets/icons/circle-check.svg',